WIP Squash me and write a real commit msg.
