+++ tests/test_filtering.py
@@ -15,13 +15,6 @@ from datetime import datetime
 from datetime import time
 from unittest2 import skip
 
-# This import is unused but is required for testing on PyPy. CPython can
-# use psycopg2, but PyPy can only use psycopg2cffi.
-try:
-    import psycopg2  # noqa
-except ImportError:
-    from psycopg2cffi import compat
-    compat.register()
 from sqlalchemy import Column
 from sqlalchemy import Date
 from sqlalchemy import DateTime
@@ -29,33 +22,14 @@ from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
 from sqlalchemy import Time
 from sqlalchemy import Unicode
-from sqlalchemy.dialects.postgresql import INET
 from sqlalchemy.ext.associationproxy import association_proxy
 from sqlalchemy.orm import backref
 from sqlalchemy.orm import relationship
-from testing.postgresql import PostgresqlFactory as PGFactory
 
 from .helpers import check_sole_error
 from .helpers import dumps
 from .helpers import loads
 from .helpers import ManagerTestBase
-
-
-#: The PostgreSQL class used to create a temporary database for testing.
-#:
-#: This class should be instantiated in the setup method of test
-#: classes, and the :class:`Postgresql.stop` method should be called on
-#: teardown.
-#:
-#: This is an optimization designed to speed up the tests that require
-#: PostgreSQL, since it can be extremely slow to initialize a PostgreSQL
-#: database before each test method.
-PostgreSQL = PGFactory(cache_initialized_db=True)
-
-
-def tearDown():
-    """Clears the cache in the :attr:`PostgreSQL` class."""
-    PostgreSQL.clear_cache()
 
 
 class SearchTestBase(ManagerTestBase):
@@ -1196,181 +1170,6 @@ class TestOperators(SearchTestBase):
         response = self.search('/api/person', filters)
         keywords = ['compare', 'value', 'NULL', 'use', 'is_null', 'operator']
         check_sole_error(response, 400, keywords)
-
-
-class TestNetworkOperators(SearchTestBase):
-    """Unit tests for the network address operators in PostgreSQL.
-
-    For more information, see `Network Address Functions and Operators`_
-    in the PostgreSQL documentation.
-
-    .. _Network Address Functions and Operators:
-       http://www.postgresql.org/docs/current/interactive/functions-net.html
-
-    """
-
-    def setUp(self):
-        super(TestNetworkOperators, self).setUp()
-
-        class Network(self.Base):
-            __tablename__ = 'network'
-            id = Column(Integer, primary_key=True)
-            address = Column(INET)
-
-        self.Network = Network
-        self.Base.metadata.create_all()
-        self.manager.create_api(Network)
-
-    def tearDown(self):
-        """Closes the database and removes the temporary directory in
-        which it lives.
-
-        """
-        super(TestNetworkOperators, self).tearDown()
-        self.database.stop()
-
-    # We know this method will be called by `setUp()` in the superclass,
-    # so we can set up the temporary database here.
-    def database_uri(self):
-        """Creates a PostgreSQL database and returns its connection URI."""
-        #: The PostgreSQL database used by the test methods in this class.
-        #:
-        #: This attribute stores a
-        #: :class:`~testing.postgresql.Postgresql` object, which must be
-        #: stopped in the :meth:`.tearDown` method.
-        self.database = PostgreSQL()
-
-        return self.database.url()
-
-    def test_is_not_equal(self):
-        """Test for the ``<>`` ("is not equal") operator.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1.5' <> inet '192.168.1.4'
-
-        """
-        network1 = self.Network(id=1, address='192.168.1.5')
-        network2 = self.Network(id=2, address='192.168.1.4')
-        self.session.add_all([network1, network2])
-        self.session.commit()
-        filters = [dict(name='address', op='<>', val='192.168.1.4')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1'] == sorted(network['id'] for network in networks)
-
-    def test_is_contained_by(self):
-        """Test for the ``<<`` ("is contained by") operator.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1.5' << inet '192.168.1/24'
-
-        """
-        network1 = self.Network(id=1, address='192.168.1.5')
-        network2 = self.Network(id=2, address='192.168.2.1')
-        self.session.add_all([network1, network2])
-        self.session.commit()
-        filters = [dict(name='address', op='<<', val='192.168.1/24')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1'] == sorted(network['id'] for network in networks)
-
-    def test_is_contained_by_or_equals(self):
-        """Test for the ``<<=`` ("is contained by or equals") operator.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1/24' <<= inet '192.168.1/24'
-
-        """
-        network1 = self.Network(id=1, address='192.168.1/24')
-        network2 = self.Network(id=2, address='192.168.1.5')
-        network3 = self.Network(id=3, address='192.168.2.1')
-        self.session.add_all([network1, network2, network3])
-        self.session.commit()
-        filters = [dict(name='address', op='<<=', val='192.168.1/24')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1', '2'] == sorted(network['id'] for network in networks)
-
-    def test_contains(self):
-        """Test for the ``>>`` ("contains") operator.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1/24' >> inet '192.168.1.5'
-
-        """
-        network1 = self.Network(id=1, address='192.168.1/24')
-        network2 = self.Network(id=2, address='192.168.2/24')
-        self.session.add_all([network1, network2])
-        self.session.commit()
-        filters = [dict(name='address', op='>>', val='192.168.1.5')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1'] == sorted(network['id'] for network in networks)
-
-    def test_contains_or_equals(self):
-        """Test for the ``>>=`` ("contains or equals") operator.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1/24' >>= inet '192.168.1/24'
-
-        """
-        network1 = self.Network(id=1, address='192.168.1/24')
-        network2 = self.Network(id=2, address='192.168/16')
-        network3 = self.Network(id=3, address='192.168.2/24')
-        self.session.add_all([network1, network2, network3])
-        self.session.commit()
-        filters = [dict(name='address', op='>>=', val='192.168.1/24')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1', '2'] == sorted(network['id'] for network in networks)
-
-    def test_contains_or_is_contained_by(self):
-        """Test for the ``&&`` ("contains or is contained by") operator.
-
-        .. warning::
-
-           This operation is only available in PostgreSQL 9.4 or later.
-
-        For example:
-
-        .. sourcecode:: postgresql
-
-           inet '192.168.1/24' && inet '192.168.1.80/28'
-
-        """
-        # network1 contains the queried subnet
-        network1 = self.Network(id=1, address='192.168.1/24')
-        # network2 is contained by the queried subnet
-        network2 = self.Network(id=2, address='192.168.1.81/28')
-        # network3 is neither
-        network3 = self.Network(id=3, address='192.168.2.1')
-        self.session.add_all([network1, network2, network3])
-        self.session.commit()
-        filters = [dict(name='address', op='&&', val='192.168.1.80/28')]
-        response = self.search('/api/network', filters)
-        document = loads(response.data)
-        networks = document['data']
-        assert ['1', '2'] == sorted(network['id'] for network in networks)
 
 
 class TestAssociationProxy(SearchTestBase):
